+++ packages/svelte2tsx/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## 0.7.61
+
+### Patch Changes
+
+-   fix: handle SvelteKit 3 changed imports for zero types feature ([#3106](https://github.com/sveltejs/language-tools/pull/3106))
+
+## 0.7.60
+
+### Patch Changes
+
+-   fix: remove type dependency on `magic-string` ([#3101](https://github.com/sveltejs/language-tools/pull/3101))
+
+## 0.7.59
+
+### Patch Changes
+
+-   fix: throw a friendly error for TypeScript 7 stable ([#3073](https://github.com/sveltejs/language-tools/pull/3073))
+
+## 0.7.58
+
+### Patch Changes
+
+-   feat: zero-config `+error.svelte` props ([#3076](https://github.com/sveltejs/language-tools/pull/3076))
+
 ## 0.7.57
 
 ### Patch Changes